requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tensorflow==2.17.0
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.25.0
numpy==1.26.4
pillow==10.4.0
pydantic==2.9.2